Bug 566618 - Avoid creating domain during wiring invalidation
There are scenarios where creating a domain will fail, for example if
the bundle file has been deleted.  There is no need to create a domain
in the case of invalidating a wiring because we are just trying to clear
the permission cache of the existing domain when invalidating a wiring.

Change-Id: Icb5d2118a267d08192f3ead21991d1955d36d34f
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>