Bug: Possible null pointer dereference of event
in org.eclipse.equinox.internal.event.EventAdminImpl.dispatchEvent(Event,
boolean)

There is a branch of statement that, if executed, guarantees that a null
value will be dereferenced, which would generate a NullPointerException
when the code is executed. Of course, the problem might be that the
branch or statement is infeasible and that the null pointer exception
can't ever be executed; deciding that is beyond the ability of SpotBugs.

Rank: Scary (6), confidence: High
Pattern: NP_NULL_ON_SOME_PATH
Type: NP, Category: CORRECTNESS (Correctness)


Change-Id: Icf310d401e255f4aea6f0c55ee021ed40e04965c
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>