Bug: org.eclipse.equinox.http.servlet.internal.context.ProxyContext$ContextAttributes is serializable but also an inner class of a non-serializable class
This Serializable class is an inner class of a non-serializable class.
Thus, attempts to serialize it will also attempt to associate instance
of the outer class with which it is associated, leading to a runtime
error.

If possible, making the inner class a static inner class should solve
the problem. Making the outer class serializable might also work, but
that would mean serializing an instance of the inner class would always
also serialize the instance of the outer class, which it often not what
you really want.

Rank: Troubling (14), confidence: High
Pattern: SE_BAD_FIELD_INNER_CLASS
Type: Se, Category: BAD_PRACTICE (Bad practice)

Change-Id: I06c93020982fd2a593189bb5650e2734410c0134
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>