Bug 570757 - Throw specific exception type if config area cannot be read
This change uses a more specific runtime exception
(IllegalStateException), if EquinoxContainer fails to initialize its
storage due to any Exception. Also the exception message is changed to
be more specific.

Previously a RuntimeException would be thrown on IOException or
BundleException, during storage initialization. This is unfortunately
not enough to cover all cases of corrupted/broken config area, as e.g. a
buffer with invalid size will result in an IllegalArgumentException.

The new exception type and message can be used to more easily identify
OSGI container initialization problems.

Change-Id: I1e6d0910550f1527b0affada0de897acbb1b8026
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>