Bug 570428 - macOS menu not clickable at launch
When running on macOS, calling the `setActivationPolicy` or
`activateWithOptions` methods on the `NSApplication` causes the menu
system to be frozen, requiring navigating away from the app and back in
again to restore.

These have been moved into the SWT's Application Delegate callback
handler for when the application has launched, which fixes the problem.
This is no longer needed in the launcher either, and showing both the
splash screen and the dialog box works as before.

Removing these calls from the launcher allows Eclipse to be launched and
have clickable menus immediately without needing to move away.

See also Bug 552063 which fixed this in SWT.

Change-Id: I6d1292b76c1a91f794ba2634e643978f5d9e5c2f
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>