Bug 570984 - StackOverflowError when Import-Package fails in fragment to circularly dependent plug-in
Using development mode in the old resolver can lead to endless recursion
when trying to set the suppliers involved in a cycle.  This is because
the cycle code is trying to drop bundles involved in the cycle that have
unsatisified requirements.  The problem is that when in development mode
the resolver allows requirements to be wired to capabilities from
unresolvable suppliers.  This means the cycle code is not properly
dropping the supplier so the supplier gets pulled into the resolve again
after dropping it which leads back to the cycle check again.

The fix is to avoid dropping the bundles involved in a cycle when in
development mode.


Change-Id: Ib3ef6b1796cd582e30a441799ed965b23cce973d
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>