Remove old build files that are never used
The various JAR files are out of date, old and never used

The ee files used to be used with the ee JARs to setup workspace
execution environments.  OSGi no longer publishes these

The bnd.bnd file was used at one point to load the framework
implementation into a bnd workspace. It is also out of date and never
used.

Change-Id: I3e4544b9654c632b1980e4300c91c1339d06cc30
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>