Bug 574432 - Test for modifying eclipse.allowAppRelaunch at runtime
This is hard to test cleanly because relaunching requires the first
application to exit, but exiting the test runner application ends the
test. Work around that by using the test runner as the second
(relaunched) application. This means that if the test fails, it shows
up not as failed but with an error "Test did not run".

Change-Id: I12dc31f42e2d7b4f261610ab0190ed5940d62d55
Signed-off-by: Christian Walther <walther@indel.ch>
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.framework/+/182604
Tested-by: Equinox Bot <equinox-bot@eclipse.org>
Reviewed-by: Thomas Watson <tjwatson@us.ibm.com>