[dogfooding][cleanup] Use new StringBuffer -> StringBuilder cleanup for tests.model
To ensure that the new JDT cleanup works correctly for clients we should
run it also on our own code base. This help to ensure that the cleanup
works fine and helps us to cleanup our own code base.

From the commit message of the cleanup:

- create a cleanup to replace usage of StringBuffer with
  StringBuilder.  The two are equivalent, but StringBuffer is
  thread-safe and synchronized which makes it slower than
  StringBuilder which was introduced in Java 1.5.


To avoid API changes, we should only use the local variable change.

Change-Id: Ie6c677c4ac8d0c3a7fa4305d2770fe5783df6473
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.bundles/+/182678
Tested-by: Equinox Bot <equinox-bot@eclipse.org>