Bug 576636 - Fix issues with added dynamic imports and multi-thread resolution
If one thread is calling
org.eclipse.osgi.internal.loader.BundleLoader.addDynamicImportPackage(ManifestElement[])
while other threads are trying to perform a resolution, dynamic or
otherwise then there is a timing issue that could prevent dynamic
resolution to the packages added with
BundleLoader.addDynamicImportPackage.  The
BundleLoader.addDynamicImportPackage gets used by WovenClassImpl to
allow weavers to add dynamic imports while they are weaving a class.

The issue is the ModuleWiring.requirements field may get overwritten by
a resolve operation in another thread doing a resolver operation after
BundleLoader.addDynamicImportPackage has been called.  This is because a
resolve operation happens on a snapshot of the wirings while holding a
read lock. Then the resolve operation happens with no locks because it
is done on copies of the live objects.  During this window, before the
resolution delta is applied, another thead could call
BundleLoader.addDynamicImportPackage which modifies the
ModuleWiring.requirements field.

Finally the resolution thread finds a resolution result and attempts to
apply it to the current wirings.  At this point it checks the timestamps
of the moduledatabase and finds it is safe to apply the results.  These
results may end up replacing the ModuleWiring.requirements of a wiring
with stale data (effectively removing the dynamically added imports.
The result is that a dynamic import resolution will then fail later.

The fix is to have the
org.eclipse.osgi.container.ModuleWiring.addDynamicImports(ModuleRevisionBuilder)
method update the ModuleDatabase timestamp such that any in progress
resolution operations can detect they are using stale data and cause
them to retry with another current snapshot.

Change-Id: I97aa78193d49d27088edb66189b2daccd29e8b06
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.framework/+/186513
Tested-by: Equinox Bot <equinox-bot@eclipse.org>