Bug 575932 - cache existence of parent directories
For dependent and registered buddy policy it is possible that a large
number of dependent and/or registered bundles can be queried for classes
for a single hosting class loader.  This results in many bundle files
getting searched.  For DirBundleFile this can be expensive because it
involves checking for the existence of each file.  This solution caches
if the parent path of the entry lookup exists.  The idea is that in most
cases where an entry lookup doesn't exists the parent directory also
doesn't exists.  We cache a smaller number of existence checks and can
do a faster lookup if files in an non-existing parent exist or not.


Change-Id: Ibc4b48800335288bb0d2a770786c612c5d5eb6dc
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.framework/+/186979
Tested-by: Equinox Bot <equinox-bot@eclipse.org>