Bug 548703 - Error dialog on start-up doesn't handle lots of input
This change adjusts the error dialog shown at Eclipse start-up, to be
able to display large amount of text. This ensures the dialog shows all
information on JVM start-up errors, since then the JVM arguments are
listed as a very long info. If the displayed message is short, the old
dialog is used.

Change-Id: I577b701d2c4bc54a54a39f3cb5b721a4279beb99
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>