Ensure lockOpen is released on unexpected errors
In CloseableBundleFile.open(boolean) the caller may request
to keep the lockOpen lock once the bundle file is successfully
opened.  Within the try block while the lockOpen is being held
during the open(boolean) call there are cases where it is
possible that the closed flag will be false upon exit if something
unexpected happens (OMM, NPE, etc).

Any time open(boolean) throws any type of throwable we need
to ensure the lockOpen lock is released.  If not then methods
like this risk exiting without ever releasing the lock:

	public BundleEntry getEntry(String path) {
		if (!lockOpen()) {
			return null;
		}
		try {
			return findEntry(path);
		} finally {
			releaseOpen();
		}
	}

Instead of handling this in open(boolean) the lock/unlock
responsibilities are moved to lockOpen() and open() methods
so we can handle the exceptional case on lockOpen() better.
Now there is an internalOpen method that must be called
while holding the openLock lock.