Improve performance for operations that modify framework state (#54)
* Avoid use of SecureRandom for creating temporary files

File.createTempFile uses SecureRandom under the covers to generate
unique file names.  This may be important for when storing files in the
global temporary file space to avoid something guessing the file and
writing to it.

The usage in the framework for createTempFile is always used for staging
file content before it gets (reliably) moved to its final destination.
This always happens as an implementation detail of storing some state in
the internal file storage of the framework.  Risk is low that some
malicious code would be able to predict the temporary file and somehow
interfere with the behavior of the framework.  Once some malicious code
has had access to the internal storage of the framework they will be
able to write to other files that have very predictable names to cause
issues.

* Limit retry attempts to MAX_TEMP_NUM (100000)

If we really have this many temporary files
there is something wrong that needs to be addressed.
Fail with IOException if we detect a 100000 attempts
to find a non-existing file