Use more Tycho-pomless: Simplify poms and delete (now) trivial poms
Remove '<skipAPIAnalysis>true</skipAPIAnalysis>' properties that are now
obsolete since the platform-build detects automatically if the API-check
should run or not.

For o.e.osgi.util the additional compiler-args are not necessary anymore
because this plug-in does not contain any sources(in that process it was
forgotten to remove them). Their removal makes the pom trivial so it can
be removed to use Tycho-pomless instead.

Part of https://github.com/eclipse-equinox/equinox/issues/9