Adding Stale bugs boot to Equinox
Keeping issue open but not reacting to them is a bad practice. With this
action a bot will comment if an issue has been inactive for 90 days to
prompt participants to take action. Then, if no additional activity
occurs after 180 days, the issue is closed.

See
https://docs.github.com/en/github-ae@latest/actions/managing-issues-and-pull-requests/closing-inactive-issues

See https://github.com/eclipse/gef-classic/pull/76#issuecomment-1158926496
for approvement