Do not track optional resources vs trigger resources separately
If the triggers are not mandatory then just resolve all unresolved
resources.
If triggers are mandatory then do not attempt to resolve all unresolved
resources (except for attaching non-payload fragments).

The way the ModuleResolver would resolve mandatory triggers followed by
trying to resolve optional resources caused issues with the timeout for
the second resolution of the optional resources.  Instead of fixing the
timeout for the second resolution this fix stops doing two resolutions
altogether.  The main difference is if the triggers are mandatory then
the resolver will only attempt to resolve the triggers and any
dependencies the trigger tries to pull in during resolution.  Before the
ModuleResolver would also attempt to resolve all unresolved resources as
optional.  In the OSGi framework this only impacts APIs (mainly on
Bundle) that require the Bundle to be resolved when the API is called.
For example, Bundle.start, Bundle.findEntries, Bundle.loadClass.