Fix console launcher not consulting eclipse.ini to read vmargs when there is no console window (#117)
Processes on Windows can be created without a window using the CREATE_NO_WINDOW flag. When the
console launcher eclipsec.exe is started as such a process, the isConsoleLauncher function wrongly
returns 0 because there is no console window and hence, only eclipsec.ini is consulted for vmargs
but not eclipse.ini (see also functions getConfigArgs and getIniFile).

The fix is to properly propagate the launcher mode from main to the launcher DLL and removing the
incomplete isConsoleLauncher function entirely.