Widen require bundle version ranges in o.e.osgi.util
o.e.osgi.util is a façade for older plugins to get access to some OSGi
packages. Plugins should be importing the OSGi package rather than
Require-Bundle on o.e.osgi.util.

Since we need to retain this façade for these older bundle, we must
not constrain the versions of the OSGi packages installed in the system.
org.osgi.util.promise has released 1.3.0 but this cannot be used with
o.e.osgi.util since it incorrectly imports too narrow a version range.

So we widen the version range up to the next major version. This will
allows the use of any compatible version of the OSGi packages in the
system.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>