Issue #255 - Improve manifest parse error for missing colon or missing space
When a manifest line has no colon ":" the parser detects the line is invalid
and assumes the content of the line is a key missing a ":" for the value.
But another cause for the invalid line is a missing space " " to indicate
a line continuation.  This change improves the error message to indicate
that a missing line continuation space may be the cause for the invalid
manifest line.