Add IPath constants EMPTY and ROOT to IPath interface
And recommend to use them instead of Path.EMPTY or Path.ROOT.

Depending on if the IPath interface or the Path class is initialized
first, Path.RUNNING_ON_WINDOWS/NO_SEGMENTS may not be initialized yet,
when it is called in the new Path(String) constructor. Therefore both
constants are moved into a static nested class, which always completes
its initialization first.