Add a standard osgi repository that includes all equinox bundles
Currently one has to wait until the next release or use the ibuilds
composite p2 repository but there is currently no way to fetch latest
equinox bundles in a more lightweight way.

This uses the new tycho-repositry-plugin to build a standard OSGi
repository that could be deployed e.g. as part of the mater verification
build.