Force English locale for date tests
When using java.text.SimpleDateFormat without a locale, the system
default locale will be considered and if that is not set to an English
variant, the "EEE" part of the date string will be in a non-English
language. To overcome this, ensure that the English locale is used for
all tests where a java.util.Date is stringified.

Signed-off-by: Torbjörn SVENSSON <torbjorn.svensson@st.com>