Add a component property type for IAdapterFactory
Component Property types are a powerful tool to leverage annotations in
combination with declarative components.

This adds a new component property type "AdapterTypes" that makes
specification of the adapted type and the adaptable types much easier
compared to the traditional approach.

With this annotation we can even default implement the
IAdapterFactory.getAdapterList() method that usually has to be kept in
sync.