Fix for TCK test ResolverHookTests.testBeginTriggers
The PR #95 changed some of the logic with the triggers
that get passed to resolver hooks.  When doing a refresh
the triggers got replaced with all unresolved revisions
when the triggers were not mandatory.  That is the case
when doing a refreshBundles call. That changed what the
framework passed to resolver hooks for triggers.

This fix saves the original triggers list so that calls
to the resolver hooks use the correct list of revisioins
for the triggers.