Proper exception in PreferencesService#setDefaultLookupOrder() #464
PreferencesService#setDefaultLookupOrder(String, String, String[]) has
recently been changed to now throwing a NullPointerException instead of
an IllegalArgumentException in case of any of its order arguments being
null. This is incompatible with the API documentation in
IPreferencesService.
This change restores the existing behavior of throwing an
IllegalArgumentException in the mentioned case.

Fixes https://github.com/eclipse-equinox/equinox/issues/464