Add a 'wires' command to allow easy investigation of bundle wires
Currently it is quite hard to find out if/why a certain bundle is using
some other bundles or are used by others.

This adds a new 'wires' command that when executed against a given
bundle prints all wires that are used by the bundle and any wires it
provides to other bundles. This results in a very powerful tool to
analyze dependency problems and find out why a bundle is actually used.

Co-authored-by: Titouan Vervack <tivervac@gmail.com>