Use submodule to include the equinox.binaries
Currently the user needs to checkout two repositories side-by-side and
specify the location of it at the commandline to fully build equinox.

This includes the other repository as a submodule now to make it easier
for users/buildscripts as they can see the content of both repositories
as one linear directory structure.