Add a command to export the module database into a file
Currently if one has a problematic resolver state it is very hard to
provide a reproducing test case for analysis.

This adds a new command exportFrameworkState that allows to export the
module database (optional with its wiring state) into a file that later
can be loaded for example in a test case to analyze it further.