Use namespace, directives and attributes to compare equal of Requirement
Currently the resolver is ultra slow in some scenarios and even never
finds a valid solution, just adding more and more permutations to check.

The reason is that one check that should actually *reduce* the number of
problematic permutations fails because it uses Requirement#equals what
is incapable of compare requirements across Resource boundaries, that
means even if the requirement describes the same restrictions it is not
considered equal and therefore selected as a possible permutation to
remove this, then the validation check detects this is unresolvable
adding possible new permutations and so on.

This now performs an equals check instead on namespace, directives and
attributes (what can be disabled by a system property) what makes the
resolver in such case from running forever to just an blink of an eye.