Remove all candidates for a substitution package that are mandatory
Currently candidates of a substitution package might be permuted even
though they are mandatory for others (e.g. they import a package that
can only be fulfilled by this one capability).

This now adds a ProblemReduction class that tries to figure out some
conflicting options and remove them from the set of items to consider.