Add example that shows possibly duplicate permutations are performed
This is using a larger set of bundles and computes a has over the dump
of the candidate set, and records if duplicates are occurring.
A duplicate is assumed if in the same session the dump text is equal to
its SHA512 hash.

Running the "manybundles" testcase shows that a total of 112655
duplicate permutations detected!