Expose the permutation data through the felix resolver logger
Currently it is not really possible to write a test that asserts
something about the complexity of a problem that the resolver has
performed. This has the risk of undetected performance regressions but
also means performance improvements can not be easily investigated. Also
in general code (like for example console commands) it could be useful
to gather some statistics.

This now adds two new methods to the logger, one that is performing a
log each time when a permutation is added or processed to either log
this by the caller or perform any statistics on them.