Add a socket event forwarder to allow external agent to watch state
Currently there is no way to watch the bundle state from the outside of
the framework, but for agents starting a framework it would be useful to
have a way to watch the bundle state. For example in PDE we launch
frameworks for different purpose in either run or debug mode but in the
UI there is only a limited visualization.
That means users are forced to either install special code (e.g. a
console) or are quite blindly needing to guess what's going on, both are
quite inconvenient as its need special interaction or can even change
the system state.

This now adds a new (internal) EventForwarder that allows to forward
events to the agent, for this purpose the framework must be started with
the property org.eclipse.osgi.internal.framework.forwarder and Equinox
will open a (local) socket connection forwarding events. This is
intentionally made internal and simple as it is not meant as a high
level debugging or management facility.