Extract retrieval of ShadowList into a method and invert conditions
Currently the code to work with the ShadowList is quite hard to
understand as it references in a quite big block and uses a negated
condition with if/else.

This extracts the retrieval of the ShadowList into a method to prevent
accidentally using a wrong reference and inverts the condition so there
is no negation involved.