Pull down the replace method from ShadowList to Candidates
The replace method is currently used on one place and the reason we
construct a new ShadowList, to remove references to this class the
method can simply pulls down to Candidates as it only operates on the
internal unmodifiable list.

This also adds a check that if the Capability to replace does not exits
it is a noop to call this method instead of running into
IndexOutOfBoundsException.