Add support for Windows on Arm64 (WoA).
A new launcher fragment for WoA is added:
'org.eclipse.equinox.launcher.win32.win32.aarch64'

Various features are also updated to include the new fragment.

To build the Equinox launcher binaries for WoA:

On a WoA box, run the following command at the root directory of this
repo:

mvn clean generate-resources -Dnative=win32.win32.aarch64

and the following launcher binaries for WoA are generated:

eclipse.exe
eclipse_11902.dll
eclipsec.exe

which are moved over to the 'equinox.binaries' repo which is
at the same directory level as that of this repo, i.e.
'..\equinox.binaries',
and to their corresponding directories there as followed:

..\equinox.binaries\org.eclipse.equinox.executable\bin\win32\win32\aarch64\eclipse.exe
..\equinox.binaries\org.eclipse.equinox.executable\bin\win32\win32\aarch64\eclipsec.exe
..\equinox.binaries\org.eclipse.equinox.launcher.win32.win32.aarch64\eclipse_11902.dll