Synchronize on saveMonitor until output is closed and done
The persistent files are not finalized until we close the
output stream to allow the ManagedOutputStream to get everything
safely persisted into the final location on disk. We should
protect multiple threads from trying to do this final step
in persisting like we do for all the writes to the output
stream.