[Build] Build native launchers on demand as part of the Jenkins pipeline
This changes the Jenkins pipeline of the equinox repository to build the
native launcher binaries automatically if the native sources have been
changed since the last launcher-build. It does the same as the existing
Jenkins free-style jobs managed only in the Jenkins-UI and triggered
manually and consequently makes them obsolete.
If the build is for the 'master' or a maintanance-branch the built
launcher binaries are pushed to the equinox.binaries repository,
choosing the correct target branch automatically.

In contrast to the previus build jobs the launcher versions are always
incremented (for the changed platforms).

Fixes https://github.com/eclipse-equinox/equinox/issues/575