Convert equinox.launcher.tests into a test-fragment
This fixes the following shortcomings associated to the launcher tests.

- When building the native launcher and executable, the entire test
project was also copied to the specialized build agents, which
unnecessarily increases their network-traffic.

- Test cannot be executed (as JUnit-Plugin test) in the IDE without
further setup.

- Setting up tests and patching a local developer eclipse should not be
done in a make-file. Its complexity should not be increased by such
task, that also only work with simple installations but not distributed
ones.