[Build] Clean-up win32/build.bat and fix its argument processing
The argument processing was fixed by removing an superfluous
'shift'-command.
This allows to use the win32/build.bat in the GH-workflow which in turn
makes it possible to remove the explicit specification of environment
variables which are set in the build.bat.

Also remove obsolete variables and machine specific paths, update
defaults listed in the explanations of the optional switches and remove
trivial labels+jumps.
Also remove specification of environment variable PROCESSOR_ARCHITECTURE
that is set by the hosting computer.