[Build] Consider JAVA_HOME for JDK headers in MacOS launcher build
and remove hard-coded paths for obsolete build-machines (OS X Yosemite
is not relevant anymore).

The value of JAVA_HOME environment variable was not considered as
location of the JDK headers and libraries against which the equinox
launcher binaries for MacOS are built.

This had the consequence that the launcher binaries for MacOS were built
against the JDK available at the build-machine and not the one defined
in the Equinox Jenkins pipeline.