Add ManifestElement.parseBundleManifest without map parameter
Currently the existing ManifestElement.parseBundleManifest with a wild
mixture of parameters, some using access restricted maps, some use null
some use a hashmap (what likely causes issue if case of headers is
different than expected).

This adds a new method that simply removes the map parameter and uses a
CaseInsensitiveDictionaryMap that is also used internally in the
framework and is most likely the most natural choice for this usecase.