[Build] Only archive just built jars and launcher binaries in Jenkins
Currently all jars in the workspace of a Jenkins build run are archived,
which also includes the entire local Maven repository and all baseline
jars. This leads to a total archive size of about 670MB (compressed).

With this only the built jars plus the native binaries are archived,
leading to a much smaller storage demands and a shorter runtime of the
archive step.