Remove unused leftovers of WPF launcher in JNIBridge
For the equinox launcher for Windows WPF the native methods
OleInitialize() and OleUninitialize() where introduced and only
implemented in the native library available for Windows WPF [1].
But during the removal of the WPF launcher [2] these methods have not
been removed.

Since the native implementation is not available anymore these native
methods cannot be used.

[1] - https://bugs.eclipse.org/bugs/show_bug.cgi?id=172367
[2] - https://bugs.eclipse.org/bugs/show_bug.cgi?id=479180