Migrate security tests to JUnit 5
The test class SignedBundleTest relies on the JUnit-3-specific
ConfigurationSessionTest. The SessionTestExtension has been introduced
as a JUnit-5-based replacement. This change adapts the test class to be
executed with JUnit 5, using the SessionTestExtension. It also adapts
other test classes in that package to JUnit 5, removing the existing
inheritance hierarchy via BaseSecurityTest, as well as the according
test suite.

# Conflicts:
#	bundles/org.eclipse.osgi.tests/src/org/eclipse/osgi/tests/security/SignedBundleTest.java