Fix module-info private packages for about_files
The framework org.eclipse.osgi cannot be used with jlink
because the module-info is not including the about_files
directory in its list of private content for the module.

The generation of the module-info from BND is not including
this folder because it is only included as a root resource
in the bundle at packaging time from the build.properties.

This change moves the about_files folder into a source folder
so that it is included in the root of the bundle just like
any other resources from the compiled source folders. This
allows BND to see the resource and generate the module-info
class correctly.