Add an equinox.api bundle
Currently we have usually have bundles that are mixed with API and
implementation, even worse those also consists of split packages, most
notable the 'org.eclipse.core.runtime' package that makes it hard to use
the API independent from the implementation and to not break existing
usages it is not easily possible to improve the situation because of the
widely use of require bundle in eclipse. Also we do not even have good
API check support for exported packages and split-packages in general
what has caused problems in the past.

This now tries to improve the situation in the following way:
-  add a new equinox.api bundle that host the actual API
- require-bundle it on the old places with reexport to ensure backward
compatibility
- add tycho-baseline-mojo to check for package version consistency
according to semantic-versioning

As a result, previous consumers of org.eclipse.equinox.common and
org.eclipse.equinox.registry can now safely use import-package instead
allowing for easy replacement of the API if desired.