Add Extender that allows bundles to declare a transformer resource
Currently it is rather inconvenient to make a transformer list available
as it requires to register an URL as service and the service itself also
needs to be registered with some properties.

This now adds an bundle extender that scans for a header
'Equinox-Transformer' where the value is pointing to a resource in the
bundle (optionally specify the transformer type). If such bundle is
found the resource is looked up and registered as a service with the
required properties so it can be discovered by the transformer.