Option to send Equinox Framework trace to the OSGi LogService
The trace messages in the framework are currently sent to the
System.out.  This commit adds an option to send the Framework
trace messages to the OSGi LogService using the LogLevel
of Trace.  The Logger name used is the trace option key.

To set Equinox trace options with LoggerAdmin the logger
name EQUINOX.TRACE is set to LogLevel TRACE. When EQUINOX.TRACE
logger is set to TRACE then all other logger names in the
LoggerContext configuration set to TRACE are also assumed to
be trace names to be enabled.