Create an SLF4j provider implementation for the Equinox Log
Currently when running an application with Equinox that uses other
logging techniques one has some kind if mixture when it comes to logging
as some parts are logging to Equinox Log (especially eclipse platform)
and some are logging to something else (e.g. jul, log4j, ...).

A common approach for this is to use bridges that redirect different
loggings to one api (e.g. slf4j) and then installing one desired
provider to use that for unified logging.

To support such deployments, this now adds a slf4-provider that acts as
such provider for slf4j so one can use it with slf4j api and also with
such bridges to get a unified logging.