Add FreeBSD platforms' build.
- part of eclipse-platform/eclipse.platform.releng.aggregator#2959
- add 2 launcher modules for FreeBSD:
  org.eclipse.equinox.launcher.gtk.freebsd.aarch64
  org.eclipse.equinox.launcher.gtk.freebsd.x86_64
- changes made to the build.sh script and make_linux.mak
(renamed to make_unix.mak) to support compiling FreeBSD native
launcher binaries
- changes made to some .c source files to support FreeBSD
- add Equinox security module for FreeBSD: org.eclipse.equinox.security.freebsd
which is just a verbatim copy of the Linux's module.

To support cross-compiling from Linux to FreeBSD, for the same hardware
architecture, a Docker container is made available in my fork with detailed
instructions:
https://github.com/chirontt/eclipse.platform.releng.aggregator/tree/master/cje-production/dockerfiles/alpine/freebsd-cross
which will be submitted to the official repo at later stage.